type-casted out small compiler-warning.
svn path=/trunk/kdelibs/; revision=35489

+++ kfile/kfileinfo.cpp
@@ -89,7 +89,7 @@ KFileInfo::KFileInfo(const KUDSEntry &e)
 	    myBaseURL = url.path( 1 ); // we want a trailing "/"
 	    break;
 	default:
-	    debug("got %ld", ( *it ).m_uds);
+	    debug("got %ld", (long int) (*it).m_uds);
 	};
     }
